Create Container and presentational parientList component
